#6 - start work on providing a mock control type
